semanticAnalysis/requirements.txt
# Core ML stack
transformers>=4.43.0
accelerate>=0.32.0
torch>=2.3.0
safetensors>=0.4.3
huggingface-hub>=0.24.0
datasets>=2.20.0
sentence-transformers>=3.0.0
pinecone-client>=3.0.0

# Optional quantization / performance
bitsandbytes; platform_system != "Windows"

# Local model serving / orchestration
ollama>=0.1.34

# Utilities
python-dotenv>=1.0.1
rich>=13.7.1
fastapi>=0.111.0
uvicorn>=0.30.0

# Testing / linting optional (comment out if not needed)
# pytest
# black
